Bring layouts into components directory
