+++ apps/smart-forms-react-app/src/components/Dashboard/DashboardHeader/DashboardHeader.tsx
@@ -1,35 +1,30 @@
-import Iconify from '../../components/Iconify';
-
-import { StyledRoot, StyledToolbar } from '../Header.styles';
-import React, { useContext } from 'react';
+import { Box, IconButton, Stack } from '@mui/material';
+import Iconify from '../../Iconify';
+import { StyledRoot, StyledToolbar } from '../../StyledComponents/Header.styles';
+import React from 'react';
 import FaceIcon from '@mui/icons-material/Face';
 import MedicalServicesIcon from '@mui/icons-material/MedicalServices';
-import AccountPopover from '../../components/Account/AccountPopover';
+import AccountPopover from '../../Account/AccountPopover';
 import { useTheme } from '@mui/material/styles';
-import PatientPopoverMenu from '../../components/Account/PatientPopoverMenu';
-import UserPopoverMenu from '../../components/Account/UserPopoverMenu';
-import UserHeader from '../../components/Account/UserHeader';
-import useResponsive from '../../custom-hooks/useResponsive';
-import Logo from '../../components/Logo/Logo';
-import { QuestionnaireProviderContext } from '../../App';
-import AssignmentIcon from '@mui/icons-material/Assignment';
-import QuestionnairePopoverMenu from '../../components/Account/QuestionnairePopoverMenu';
-import { Box, IconButton, Stack, Typography } from '@mui/material';
+import PatientPopoverMenu from '../../Account/PatientPopoverMenu';
+import UserPopoverMenu from '../../Account/UserPopoverMenu';
+import UserHeader from '../../Account/UserHeader';
+import useResponsive from '../../../custom-hooks/useResponsive';
+import Logo from '../../Logo/Logo';
 
 interface Props {
   onOpenNav: () => void;
 }
 
-function RendererHeader(props: Props) {
+function DashboardHeader(props: Props) {
   const { onOpenNav } = props;
 
   const theme = useTheme();
 
   const isDesktop = useResponsive('up', 'lg');
-  const { questionnaire } = useContext(QuestionnaireProviderContext);
 
   return (
-    <StyledRoot sx={{ boxShadow: theme.customShadows.z4 }}>
+    <StyledRoot>
       <StyledToolbar>
         <IconButton
           onClick={onOpenNav}
@@ -46,14 +41,6 @@ function RendererHeader(props: Props) {
             <Logo />
           </Box>
         ) : null}
-
-        {isDesktop ? (
-          <Box sx={{ px: 1 }}>
-            <Typography variant="subtitle1" color="text.primary">
-              {questionnaire.title}
-            </Typography>
-          </Box>
-        ) : null}
         <Box sx={{ flexGrow: 1 }} />
 
         <Stack
@@ -68,11 +55,6 @@ function RendererHeader(props: Props) {
             <UserHeader />
           ) : (
             <>
-              <AccountPopover
-                bgColor={theme.palette.primary.main}
-                displayIcon={<AssignmentIcon sx={{ color: theme.palette.common.white }} />}
-                menuContent={<QuestionnairePopoverMenu />}
-              />
               <AccountPopover
                 bgColor={theme.palette.secondary.main}
                 displayIcon={<FaceIcon sx={{ color: theme.palette.common.white }} />}
@@ -91,4 +73,4 @@ function RendererHeader(props: Props) {
   );
 }
 
-export default RendererHeader;
+export default DashboardHeader;
