+++ apps/smart-forms-react-app/src/components/Dashboard/DashboardLayout.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
-import DashboardHeader from './DashboardHeader';
-import Nav from './nav/Nav';
+import DashboardHeader from './DashboardHeader/DashboardHeader';
+import DashboardNav from './DashboardNav/DashboardNav';
 import { Main, StyledRoot } from './DashboardLayout.styles';
 import { Outlet } from 'react-router-dom';
 import SelectedQuestionnaireContextProvider from '../../custom-contexts/SelectedQuestionnaireContext';
@@ -11,7 +11,7 @@ function DashboardLayout() {
   return (
     <StyledRoot>
       <DashboardHeader onOpenNav={() => setOpen(true)} />
-      <Nav openNav={open} onCloseNav={() => setOpen(false)} />
+      <DashboardNav openNav={open} onCloseNav={() => setOpen(false)} />
 
       <Main>
         <SelectedQuestionnaireContextProvider>
