+++ apps/smart-forms-react-app/src/components/DebugFooter/DebugFooter.styles.js
