+++ apps/smart-forms-react-app/src/components/Operations/SaveAsFinalOperation.tsx
@@ -1,11 +1,11 @@
 import TaskAltIcon from '@mui/icons-material/TaskAlt';
 import React, { useContext, useState } from 'react';
-import { removeHiddenAnswers, saveQuestionnaireResponse } from '../functions/SaveQrFunctions';
-import { LaunchContext } from '../custom-contexts/LaunchContext';
-import { QuestionnaireProviderContext, QuestionnaireResponseProviderContext } from '../App';
-import { RendererContext } from '../layouts/renderer/RendererLayout';
-import { EnableWhenContext } from '../custom-contexts/EnableWhenContext';
-import { OperationItem } from '../components/Nav/ViewerNav/OperationSection';
+import { removeHiddenAnswers, saveQuestionnaireResponse } from '../../functions/SaveQrFunctions';
+import { LaunchContext } from '../../custom-contexts/LaunchContext';
+import { QuestionnaireProviderContext, QuestionnaireResponseProviderContext } from '../../App';
+import { RendererContext } from '../Renderer/RendererLayout';
+import { EnableWhenContext } from '../../custom-contexts/EnableWhenContext';
+import { OperationItem } from '../Nav/ViewerNav/OperationSection';
 import {
   Button,
   CircularProgress,
